+++ src/server/services/drizzle/config.ts
@@ -1,0 +1,3 @@
+export const config = {
+  connectionString: process.env.DATABASE_URL
+}
